refactor(components): убрать лишний компонент обертку MainBlock
