+++ src/App.js
@@ -1,11 +1,14 @@
 import React, { Component } from 'react';
 import './App.css';
 import Header from './components/Header/Header';
-import MainBlock from './components/Mainblock/Mainblock';
+import Login from './components/Login/Login';
+import MapBlock from './components/MapBlock/MapBlock';
+import Profile from './components/Profile/Profile';
+import Signup from './components/SignUp/Signup';
 
 class App extends Component {
   state = {
-    currentTab: "login",
+    currentTab: "signup",
     paymentData: false
   };
 
@@ -19,27 +22,26 @@ class App extends Component {
     this.setState({
       paymentData: true
     })
-  }
+  };
 
   render () {
     // console.log('App state: ', this.state);
+    const tabs = {
+      signup: <Signup handleChangeCurrentTab={ this.handleChangeCurrentTab } />,
+      login: <Login handleChangeCurrentTab={ this.handleChangeCurrentTab } />,
+      profile: <Profile handleChangeCurrentTab={ this.handleChangeCurrentTab } handleChangePaymentData={ this.handleChangePaymentData } />,
+      mapblock: <MapBlock handleChangeCurrentTab={ this.handleChangeCurrentTab }  paymentData={ this.state.paymentData } />
+    };
+
     return (
       <div className="app">
-        { this.state["currentTab"] !== "login" && <Header handleChangeCurrentTab={ this.handleChangeCurrentTab } /> }
+        { this.state["currentTab"] !== "login" && this.state["currentTab"] !== "signup" && <Header handleChangeCurrentTab={ this.handleChangeCurrentTab } /> }
         <div className="main-block">
-          <MainBlock 
-            currnetTab={ this.state.currentTab }
-            paymentData={ this.state.paymentData }
-            handleChangeCurrentTab={ this.handleChangeCurrentTab }
-            handleChangePaymentData={ this.handleChangePaymentData } />
+          {tabs[this.state["currentTab"]]}
         </div>
       </div>
     );
   }
 }
-
-// mainblock удалить и впихнуть все в App через объект с ключами
-// отдельно сделать компоненту входа и компоненту регистрации
-
 
 export default App;
