+++ src/components/Header/Header.js
@@ -1,20 +1,25 @@
 import React from 'react';
 import HeaderMod from './Header.module.css';
 import logo from './logo.svg';
+import PropTypes from 'prop-types';
 
-const Header = (props) => {
+const Header = ({ handleChangeCurrentTab }) => {
    return (
       <header className={ HeaderMod.app_header }>
          <img src={logo} className={ HeaderMod.logo } alt="logo" />
          <div className={ HeaderMod.menu }>
             <ul>
-               <li onClick={ () => props.handleChangeCurrentTab("mapblock") }>Карта</li>
-               <li onClick={ () => props.handleChangeCurrentTab("profile") }>Профиль</li>
-               <li onClick={ () => props.handleChangeCurrentTab("login") }>Выйти</li>
+               <li onClick={ () => handleChangeCurrentTab("mapblock") }>Карта</li>
+               <li onClick={ () => handleChangeCurrentTab("profile") }>Профиль</li>
+               <li onClick={ () => handleChangeCurrentTab("login") }>Выйти</li>
             </ul>
          </div>
       </header>
    )
+}
+
+Header.propTypes = {
+   handleChangeCurrentTab: PropTypes.func
 }
 
 export default Header;
