+++ src/components/Login/Login.module.css
@@ -3,15 +3,20 @@
    align-items: center;
    padding-top: 68px;
    height: calc(100vh - 68px);
-   background: url(../../../main-bg.png) center 50% no-repeat;
+   background: url(../../main-bg.png) center 50% no-repeat;
 }
 
 .login > div {
    width: 50%;
 }
 
-.logo {
+.logo_block {
    text-align: right;
+}
+
+.logo {
+   margin-right: 160px;
+   width: 156px;
 }
 
 .popup > div {
@@ -20,4 +25,24 @@
    background-color: #fff;
    padding: 45px;
    border-radius: 5px;
+}
+
+.log_in h2 {
+   margin-top: 0;
+}
+
+.log_in span {
+   text-decoration: none;
+   color: #1473e6;
+   cursor: pointer;
+   transition: all .3s;
+}
+
+.log_in span:hover {
+   text-decoration: underline;
+   transition: all .3s;
+}
+
+.log_in input[type="submit"] {
+   width: 160px;
 }
