--- src/components/Mainblock/Login/Logo/Logo.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import LogoMod from './Logo.module.css';
-import logo from './logo.svg';
-
-const Logo = () => {
-   return (
-      <img src={logo} className={ LogoMod.logo } alt="logo" />
-   );
-}
-
-export default Logo;
