--- src/components/Mainblock/Login/Logo/Logo.module.css
@@ -1,4 +1,0 @@
-.logo {
-   margin-right: 160px;
-   width: 156px;
-}
