--- src/components/Mainblock/Login/Popup/LoginPopup/LoginPopup.module.css
@@ -1,19 +1,0 @@
-.log_in h2 {
-   margin-top: 0;
-}
-
-.log_in span {
-   text-decoration: none;
-   color: #1473e6;
-   cursor: pointer;
-   transition: all .3s;
-}
-
-.log_in span:hover {
-   text-decoration: underline;
-   transition: all .3s;
-}
-
-.log_in input[type="submit"] {
-   width: 160px;
-}
