--- src/components/Mainblock/Login/Popup/SigninPopup/SigninPopup.module.css
@@ -1,15 +1,0 @@
-.sign_in h2 {
-   margin-top: 0;
-}
-
-.sign_in span {
-   text-decoration: none;
-   color: #1473e6;
-   cursor: pointer;
-   transition: all .3s;
-}
-
-.sign_in span:hover {
-   text-decoration: underline;
-   transition: all .3s;
-}
