+++ src/components/SignUp/Signup.module.css
@@ -1,0 +1,44 @@
+.login {
+   display: flex;
+   align-items: center;
+   padding-top: 68px;
+   height: calc(100vh - 68px);
+   background: url(../../main-bg.png) center 50% no-repeat;
+}
+
+.login > div {
+   width: 50%;
+}
+
+.logo_block {
+   text-align: right;
+}
+
+.logo {
+   margin-right: 160px;
+   width: 156px;
+}
+
+.popup > div {
+   width: 70%; /* если размер окна меньше 1100px, то уменьшаем размер высплывающего окна или меняем направление главной оси родительского блока flex-direction: column; */
+   max-width: 350px;
+   background-color: #fff;
+   padding: 45px;
+   border-radius: 5px;
+}
+
+.sign_in h2 {
+   margin-top: 0;
+}
+
+.sign_in span {
+   text-decoration: none;
+   color: #1473e6;
+   cursor: pointer;
+   transition: all .3s;
+}
+
+.sign_in span:hover {
+   text-decoration: underline;
+   transition: all .3s;
+}
